add support for dns-01 challenge - try 2

+++ letsencrypt.sh
@@ -33,6 +33,7 @@ load_config() {
   # Default values
   CA="https://acme-v01.api.letsencrypt.org/directory"
   LICENSE="https://letsencrypt.org/documents/LE-SA-v1.0.1-July-27-2015.pdf"
+  CHALLENGETYPE="http-01"  
   HOOK=
   RENEW_DAYS="30"
   PRIVATE_KEY="${BASEDIR}/private_key.pem"
@@ -64,6 +65,20 @@ load_config() {
 
   if [[ -n "${PARAM_HOOK:-}" ]]; then
     HOOK="${PARAM_HOOK}"
+  fi
+
+  if [[ -n "${PARAM_CHALLENGETYPE:-}" ]]; then
+    CHALLENGETYPE="${PARAM_CHALLENGETYPE}"
+  fi
+
+  case "${CHALLENGETYPE}" in
+    http-01|dns-01) ;; # We suppport these types
+    *) echo "Unknown challenge type ${CHALLENGETYPE} ... can not continue"; exit 1;;
+  esac
+
+  if [[ "${CHALLENGETYPE}" = "dns-01" ]] && [[ -z "${HOOK}" ]]; then
+   echo "Challenge type dns-01 needs a hook script for deployment ... can not continue"
+   exit 1
   fi
 }
 
@@ -270,7 +285,7 @@ sign_domain() {
 
     challenges="$(printf '%s\n' "${response}" | grep -Eo '"challenges":[^\[]*\[[^]]*]')"
     repl=$'\n''{' # fix syntax highlighting in Vim
-    challenge="$(printf "%s" "${challenges//\{/${repl}}" | grep 'http-01')"
+    challenge="$(printf "%s" "${challenges//\{/${repl}}" | grep \""${CHALLENGETYPE}"\")"
     challenge_token="$(printf '%s' "${challenge}" | get_json_string_value token | sed 's/[^A-Za-z0-9_\-]/_/g')"
     challenge_uri="$(printf '%s' "${challenge}" | get_json_string_value uri)"
 
@@ -281,12 +296,19 @@ sign_domain() {
     # Challenge response consists of the challenge token and the thumbprint of our public certificate
     keyauth="${challenge_token}.${thumbprint}"
 
-    # Store challenge response in well-known location and make world-readable (so that a webserver can access it)
-    printf '%s' "${keyauth}" > "${WELLKNOWN}/${challenge_token}"
-    chmod a+r "${WELLKNOWN}/${challenge_token}"
+    if [[ "${CHALLENGETYPE}" = "http-01" ]]; then
+      # Store challenge response in well-known location and make world-readable (so that a webserver can access it)
+      printf '%s' "${keyauth}" > "${WELLKNOWN}/${challenge_token}"
+      chmod a+r "${WELLKNOWN}/${challenge_token}"
+    fi
+
+    keyauth_hook="${keyauth}"
+    if [[ "${CHALLENGETYPE}" = "dns-01" ]]; then
+      keyauth_hook="$(printf '%s' "${keyauth}" | openssl sha -sha256 -binary | urlbase64)"
+    fi
 
     # Wait for hook script to deploy the challenge if used
-    [[ -n "${HOOK}" ]] && ${HOOK} "deploy_challenge" "${altname}" "${challenge_token}" "${keyauth}"
+    [[ -n "${HOOK}" ]] && ${HOOK} "deploy_challenge" "${altname}" "${challenge_token}" "${keyauth_hook}"
 
     # Ask the acme-server to verify our challenge and wait until it is no longer pending
     echo " + Responding to challenge for ${altname}..."
@@ -299,11 +321,13 @@ sign_domain() {
       status="$(http_request get "${challenge_uri}" | get_json_string_value status)"
     done
 
-    rm -f "${WELLKNOWN}/${challenge_token}"
+    if [[ "${CHALLENGETYPE}" = "http-01" ]]; then
+      rm -f "${WELLKNOWN}/${challenge_token}"
+    fi
 
     # Wait for hook script to clean the challenge if used
     if [[ -n "${HOOK}" ]] && [[ -n "${challenge_token}" ]]; then
-      ${HOOK} "clean_challenge" "${altname}" "${challenge_token}" "${keyauth}"
+      ${HOOK} "clean_challenge" "${altname}" "${challenge_token}" "${keyauth_hook}"
     fi
 
     if [[ "${status}" = "valid" ]]; then
@@ -479,7 +503,7 @@ command_help() {
 command_env() {
   echo "# letsencrypt.sh configuration"
   load_config
-  typeset -p CA LICENSE HOOK RENEW_DAYS PRIVATE_KEY KEYSIZE WELLKNOWN PRIVATE_KEY_RENEW OPENSSL_CNF CONTACT_EMAIL LOCKFILE
+  typeset -p CA LICENSE CHALLENGETYPE HOOK RENEW_DAYS PRIVATE_KEY KEYSIZE WELLKNOWN PRIVATE_KEY_RENEW OPENSSL_CNF CONTACT_EMAIL LOCKFILE
 }
 
 # Main method (parses script arguments and calls command_* methods)
@@ -564,6 +588,14 @@ main() {
         shift 1
         check_parameters "${1:-}"
         PARAM_HOOK="${1}"
+        ;;
+
+      # PARAM_Usage: --challenge (-t) http-01|dns-01
+      # PARAM_Description: Which challenge should be used? Currently http-01 and dns-01 are supported
+      --challenge|-t)
+        shift 1
+        check_parameters "${1:-}"
+        PARAM_CHALLENGETYPE="${1}"
         ;;
 
       *)
